renamed interface and default impl
