+++ src/test/scala/beam/agentsim/agents/ridehail/graph/GraphRunHelper.scala
@@ -1,5 +1,5 @@
 package beam.agentsim.agents.ridehail.graph
-import beam.router.r5.NetworkCoordinator
+import beam.router.r5.DefaultNetworkCoordinator
 import beam.sim.{BeamHelper, BeamServices}
 import beam.sim.config.{BeamConfig, MatSimBeamConfigBuilder}
 import beam.sim.population.DefaultPopulationAdjustment
@@ -23,7 +23,7 @@ class GraphRunHelper(childModule: AbstractModule, baseConfig: Config) extends Be
   matsimConfig.planCalcScore().setMemorizingExperiencedPlans(true)
   FileUtils.setConfigOutputFile(beamConfig, matsimConfig)
 
-  private val networkCoordinator = new NetworkCoordinator(beamConfig)
+  private val networkCoordinator = new DefaultNetworkCoordinator(beamConfig)
   networkCoordinator.loadNetwork()
   networkCoordinator.convertFrequenciesToTrips()
