+++ src/test/scala/beam/integration/DriveTransitSpec.scala
@@ -1,6 +1,6 @@
 package beam.integration
 
-import beam.router.r5.NetworkCoordinator
+import beam.router.r5.DefaultNetworkCoordinator
 import beam.sim.config.{BeamConfig, MatSimBeamConfigBuilder}
 import beam.sim.population.DefaultPopulationAdjustment
 import beam.sim.{BeamHelper, BeamServices}
@@ -49,7 +49,7 @@ class DriveTransitSpec extends WordSpecLike with Matchers with BeamHelper {
       FileUtils.setConfigOutputFile(beamConfig, matsimConfig)
       val scenario =
         ScenarioUtils.loadScenario(matsimConfig).asInstanceOf[MutableScenario]
-      val networkCoordinator = new NetworkCoordinator(beamConfig)
+      val networkCoordinator = new DefaultNetworkCoordinator(beamConfig)
       networkCoordinator.loadNetwork()
       scenario.setNetwork(networkCoordinator.network)
       var nDepartures = 0
