+++ src/test/scala/beam/integration/ridehail/RideHailReplaceAllocationSpec.scala
@@ -1,7 +1,7 @@
 package beam.integration.ridehail
 
 import beam.agentsim.agents.ridehail.allocation.RideHailResourceAllocationManager
-import beam.router.r5.NetworkCoordinator
+import beam.router.r5.DefaultNetworkCoordinator
 import beam.sim.{BeamHelper, BeamServices}
 import beam.sim.config.BeamConfig
 import beam.sim.population.DefaultPopulationAdjustment
@@ -26,7 +26,7 @@ class RideHailReplaceAllocationSpec extends FlatSpec with BeamHelper with Mockit
     val beamConfig = BeamConfig(config)
     FileUtils.setConfigOutputFile(beamConfig, matsimConfig)
 
-    val networkCoordinator = new NetworkCoordinator(beamConfig)
+    val networkCoordinator = new DefaultNetworkCoordinator(beamConfig)
     networkCoordinator.loadNetwork()
     networkCoordinator.convertFrequenciesToTrips()
